requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
pymongo==4.6.1

# HTTP & API
requests==2.31.0
schedule==1.1.0

# Configuration
python-dotenv==1.0.0
dnspython==2.4.2

# Production Server
gunicorn==21.2.0
waitress==3.0.0

# Rate Limiting & Caching
Flask-Limiter==3.5.0
Flask-Caching==2.1.0

# Security
Flask-Talisman==1.1.0

# Monitoring & Logging
python-json-logger==2.0.7

# Data Export
pandas==2.1.4
openpyxl==3.1.2

# API Documentation
flasgger==0.9.7.1
